requirements.txt
appnope==0.1.0
attrs==17.4.0
backports-abc==0.5
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
biopython==1.70
bleach==1.5.0
certifi==2018.8.24
chardet==3.0.4
configparser==3.5.0
cycler==0.10.0
decorator==4.1.2
entrypoints==0.2.3
enum34==1.1.6
FlowCal==1.1.4
funcsigs==1.0.2
functools32==3.2.3.post2
goatools==0.8.2
html5lib==0.9999999
idna==2.7
ipykernel==4.6.1
ipython==5.3.0
ipython-genutils==0.2.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-core==4.3.0
MarkupSafe==1.0
matplotlib==2.1.0
mistune==0.7.4
more-itertools==4.1.0
nbconvert==5.2.1
nbformat==4.4.0
nose==1.3.7
notebook==5.0.0
numexpr==2.6.4
numpy==1.13.3
packaging==16.8
palettable==3.1.0
pandas==0.21.0
pandocfilters==1.4.2
pathlib2==2.3.0
patsy==0.5.0
pexpect==4.2.1
pickleshare==0.7.4
plotly==3.2.0
pluggy==0.6.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.5.3
pyaml==17.10.0
pybedtools==0.7.10
pydot==1.2.4
Pygments==2.2.0
pyliftover==0.3
pyparsing==2.2.0
pysam==0.13
pytest==3.5.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
pyzmq==16.0.2
requests==2.19.1
retrying==1.3.3
roman==3.0
scandir==1.5
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
statsmodels==0.8.0
subprocess32==3.2.7
tables==3.4.2
terminado==0.6
testpath==0.3
tornado==4.5.2
traitlets==4.3.2
uncompyle2==1.1
urllib3==1.23
wcwidth==0.1.7
wget==3.2
xlrd==1.1.0
XlsxWriter==1.0.2
